Move chart data related modules under their own dir
This file structure is starting to make more sense, bit by bit...
